refactor(game_screen): extract GameListView into its own file
Move GameListView and its State out of the 3,693-line my_games_list.dart
into a dedicated lib/screens/game_screen/game_list_view.dart (~443 lines),
shrinking my_games_list.dart by ~430 lines with zero behaviour change.

The State class is renamed _GameListViewState -> public GameListViewState
so the parent's GlobalKey<GameListViewState> can reference it across files
(standard Flutter pattern for GlobalKey-exposed state). Four imports that
were only used by the moved class are dropped from my_games_list.dart and
carried into the new file.

Pure move: the class body is byte-identical to the original aside from the
state rename and dart-format normalisation of previously-unformatted lines.
Verified: flutter analyze introduces no new issues, dart format clean,
flutter test 198/198 pass.
